services/dashboard/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
python-dotenv
transformers
torch